Move SetupTestStateDiffIndexer to shared test helpers
