+++ .github/workflows/on-pr.yaml
@@ -64,7 +64,7 @@ jobs:
       DB_WRITE: true
       ETH_FORWARD_ETH_CALLS: false
       ETH_PROXY_ON_ERROR: false
-      ETH_HTTP_PATH: ""
+      ETH_HTTP_PATH: "dapptools:8545"
     strategy:
       matrix:
         go-version: [1.16.x]
@@ -122,7 +122,7 @@ jobs:
       DB_WRITE: true
       ETH_FORWARD_ETH_CALLS: false
       ETH_PROXY_ON_ERROR: false
-      ETH_HTTP_PATH: ""
+      ETH_HTTP_PATH: "dapptools:8545"
       WATCHED_ADDRESS_GAP_FILLER_ENABLED: true
       WATCHED_ADDRESS_GAP_FILLER_INTERVAL: 2
     strategy:
